Update Dockerfile
Download xnat installation manually and install offline.   
The downloaded .war file is supposed to be placed with the Dockerfile.  
It will be unzipped by docker's `ADD` to the original directory.
The filename should be consistent with what is specified in the file. Usually, the name of the official release is OK.
Also note that xnat version should be consistent.